aded wrap-up script
The « summary2csv » script is there to read simulation data and wrap it
up in a cvs files (one with all the tested parameterizations, the other
with only those that reach the maximum score of 14).
Reads the content of the local directories beginning with « 2017 »
(HACK), finds the score and the parameters inside, extract them and
writes them in the cvs files.